# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feed:
    get:
      tags: [posts]
      operationId: getFeed
      summary: Get paginated post feed
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Feed posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedResponse'
  /posts:
    post:
      tags: [posts]
      operationId: createPost
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostBody'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostRecord'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /posts/{postId}/like:
    post:
      tags: [social]
      operationId: likePost
      summary: Like a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LikeResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /posts/{postId}/save:
    post:
      tags: [social]
      operationId: savePost
      summary: Save/unsave a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Save toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /posts/{postId}/comments:
    get:
      tags: [posts]
      operationId: getPostComments
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comments list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentsResponse'
    post:
      tags: [posts]
      operationId: addComment
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCommentBody'
      responses:
        '201':
          description: Comment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentRecord'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /notifications:
    get:
      tags: [notifications]
      operationId: getNotifications
      summary: Get notifications for current user
      responses:
        '200':
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /notifications/read-all:
    post:
      tags: [notifications]
      operationId: markAllNotificationsRead
      summary: Mark all notifications as read
      responses:
        '200':
          description: All marked read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /messages/threads:
    get:
      tags: [messages]
      operationId: getMessageThreads
      summary: Get all message threads for current user
      responses:
        '200':
          description: Threads list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /messages/threads/{threadId}:
    get:
      tags: [messages]
      operationId: getThreadMessages
      summary: Get messages in a thread
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /messages/send:
    post:
      tags: [messages]
      operationId: sendMessage
      summary: Send a message to a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageBody'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageRecord'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
...
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
...
    PostRecord:
      type: object
      required: [id, authorId, authorName, caption, likeCount, commentCount, saveCount, viewCount, isPatronOnly, createdAt]
      properties:
        id:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        authorAvatarUrl:
          type: [string, 'null']
        videoUrl:
          type: [string, 'null']
        thumbnailUrl:
          type: [string, 'null']
        caption:
          type: string
        technique:
          type: [string, 'null']
        medium:
          type: [string, 'null']
        tags:
          type: array
          items:
            type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        saveCount:
          type: integer
        viewCount:
          type: integer
        isPatronOnly:
          type: boolean
        craftScore:
          type: [integer, 'null']
        isLiked:
          type: boolean
        isSaved:
          type: boolean
        createdAt:
          type: string
    FeedResponse:
      type: object
      required: [posts, hasMore]
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/PostRecord'
        hasMore:
          type: boolean
        nextCursor:
          type: [string, 'null']
    CreatePostBody:
      type: object
      required: [caption]
      properties:
        caption:
          type: string
        videoUrl:
          type: string
        thumbnailUrl:
          type: string
        technique:
          type: string
        medium:
          type: string
        tags:
          type: array
          items:
            type: string
        isPatronOnly:
          type: boolean
    LikeResponse:
      type: object
      required: [liked, likeCount]
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
    SaveResponse:
      type: object
      required: [saved, saveCount]
      properties:
        saved:
          type: boolean
        saveCount:
          type: integer
...
    CommentRecord:
      type: object
      required: [id, postId, authorId, authorName, text, likeCount, createdAt]
      properties:
        id:
          type: string
        postId:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        authorAvatarUrl:
          type: [string, 'null']
        text:
          type: string
        likeCount:
          type: integer
        createdAt:
          type: string
    CommentsResponse:
      type: object
      required: [comments]
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentRecord'
    AddCommentBody:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
...
    NotificationRecord:
      type: object
      required: [id, userId, type, text, read, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
        fromId:
          type: [string, 'null']
        fromName:
          type: [string, 'null']
        fromAvatarUrl:
          type: [string, 'null']
        text:
          type: string
        link:
          type: [string, 'null']
        read:
          type: boolean
        createdAt:
          type: string
    NotificationsResponse:
      type: object
      required: [notifications]
      properties:
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/NotificationRecord'
    ThreadRecord:
      type: object
      required: [id, participantA, participantB, lastMessageAt]
      properties:
        id:
          type: string
        participantA:
          type: string
        participantB:
          type: string
        lastMessageAt:
          type: string
        lastMessageText:
          type: [string, 'null']
        otherUserId:
          type: string
        otherUserName:
          type: [string, 'null']
        otherUserAvatar:
          type: [string, 'null']
        unreadCount:
          type: integer
    ThreadsResponse:
      type: object
      required: [threads]
      properties:
        threads:
          type: array
          items:
            $ref: '#/components/schemas/ThreadRecord'
    MessageRecord:
      type: object
      required: [id, threadId, senderId, senderName, text, read, createdAt]
      properties:
        id:
          type: string
        threadId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        senderAvatarUrl:
          type: [string, 'null']
        text:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
    MessagesResponse:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/MessageRecord'
    SendMessageBody:
      type: object
      required: [recipientId, text]
      properties:
        recipientId:
          type: string
        text:
          type: string
          minLength: 1